feat: Add Makefile for building and running the program
This commit adds a Makefile to the project, which allows for easier building and running of the program. The Makefile includes targets for building the executable, running the program, and cleaning the build artifacts. This provides an alternative method to running the program using `go run`. The Makefile also accepts a `cep` argument to specify the zip code to search for. This change improves the development workflow and makes it more convenient to work with the project.